# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AIAnalysisResponse": {
        "description": "返回最近一次结构化 AI 分析及其数据溯源。",
        "properties": {
          "actions": {
            "items": {
              "type": "string"
            },
            "title": "Actions",
            "type": "array"
          },
          "conclusion": {
            "$ref": "#/components/schemas/AnalysisConclusion"
          },
          "dataAsOf": {
            "title": "Dataasof",
            "type": "string"
          },
          "dataSources": {
            "items": {
              "type": "string"
            },
            "title": "Datasources",
            "type": "array"
          },
          "disclaimer": {
            "default": "仅供个人研究参考，不构成投资建议。",
            "title": "Disclaimer",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instrument": {
            "$ref": "#/components/schemas/AnalysisInstrumentResponse"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/AnalysisMetricResponse"
            },
            "title": "Metrics",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "title": "Risks",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "instrument",
          "provider",
          "model",
          "dataAsOf",
          "dataSources",
          "metrics",
          "conclusion",
          "summary",
          "highlights",
          "risks",
          "actions",
          "generatedAt"
        ],
        "title": "AIAnalysisResponse",
        "type": "object"
      },
      "AIConversationMessageCreate": {
        "additionalProperties": false,
        "description": "校验用户发送给 Agent 的单轮文本。",
        "properties": {
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "AIConversationMessageCreate",
        "type": "object"
      },
      "AIConversationMessageResponse": {
        "description": "返回一条已保存或正在生成的对话消息。",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/AIMessageRole"
          },
          "status": {
            "$ref": "#/components/schemas/AIMessageStatus"
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "content",
          "createdAt"
        ],
        "title": "AIConversationMessageResponse",
        "type": "object"
      },
      "AIConversationResponse": {
        "description": "返回当前用户独立的 Codex 会话及完整消息。",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instrument": {
            "$ref": "#/components/schemas/AnalysisInstrumentResponse"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/AIConversationMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AIConversationStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "instrument",
          "title",
          "status",
          "messages",
          "createdAt",
          "updatedAt"
        ],
        "title": "AIConversationResponse",
        "type": "object"
      },
      "AIConversationStatus": {
        "description": "表示一个会话当前是否可接受下一轮消息。",
        "enum": [
          "IDLE",
          "RUNNING",
          "FAILED"
        ],
        "title": "AIConversationStatus",
        "type": "string"
      },
      "AIMessageRole": {
        "description": "区分用户输入和 Agent 回复。",
        "enum": [
          "USER",
          "ASSISTANT"
        ],
        "title": "AIMessageRole",
        "type": "string"
      },
      "AIMessageStatus": {
        "description": "记录单条消息是否仍在生成或已经结束。",
        "enum": [
          "STREAMING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "AIMessageStatus",
        "type": "string"
      },
      "AnalysisConclusion": {
        "description": "用中性研究结论替代确定性买卖指令。",
        "enum": [
          "POSITIVE",
          "NEUTRAL",
          "CAUTIOUS",
          "INSUFFICIENT_DATA"
        ],
        "title": "AnalysisConclusion",
        "type": "string"
      },
      "AnalysisInstrumentResponse": {
        "description": "返回分析结果对应的稳定资产身份。",
        "properties": {
          "assetType": {
            "$ref": "#/components/schemas/AssetType"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "market": {
            "$ref": "#/components/schemas/Market"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "id",
          "assetType",
          "market",
          "exchange",
          "ticker",
          "name",
          "currency"
        ],
        "title": "AnalysisInstrumentResponse",
        "type": "object"
      },
      "AnalysisMetricResponse": {
        "description": "返回页面直接展示的一项已计算指标。",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "title": "AnalysisMetricResponse",
        "type": "object"
      },
      "AssetType": {
        "description": "区分一期支持的股票与基金。",
        "enum": [
          "STOCK",
          "FUND"
        ],
        "title": "AssetType",
        "type": "string"
      },
...
      "Currency": {
        "description": "表示权威财务值的币种。",
        "enum": [
          "CNY"
        ],
        "title": "Currency",
        "type": "string"
      },
      "DataFreshness": {
        "description": "表示本地数据是否仍在约定的新鲜时间窗口内。",
        "enum": [
          "FRESH",
          "STALE"
        ],
        "title": "DataFreshness",
        "type": "string"
      },
...
      "Exchange": {
        "description": "区分 A 股交易所与中国公募基金登记域。",
        "enum": [
          "SSE",
          "SZSE",
          "BSE",
          "FUND_CN"
        ],
        "title": "Exchange",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstrumentListResponse": {
        "description": "返回稳定分页的资产搜索结果。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstrumentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "title": "InstrumentListResponse",
        "type": "object"
      },
      "InstrumentResponse": {
        "description": "返回一期资产主数据和互不混淆的最新价格列表。",
        "properties": {
          "assetType": {
            "$ref": "#/components/schemas/AssetType"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latestPrices": {
            "items": {
              "$ref": "#/components/schemas/LatestPriceResponse"
            },
            "title": "Latestprices",
            "type": "array"
          },
          "market": {
            "$ref": "#/components/schemas/Market"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "sourceUpdatedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceupdatedat"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "assetType",
          "market",
          "exchange",
          "ticker",
          "name",
          "currency",
          "source",
          "sourceUpdatedAt",
          "updatedAt",
          "latestPrices"
        ],
        "title": "InstrumentResponse",
        "type": "object"
      },
      "LatestPriceResponse": {
        "description": "返回一种明确口径的最新价格及其业务时间和抓取时间。",
        "properties": {
          "asOfAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asofat"
          },
          "asOfDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asofdate"
          },
          "changeRate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changerate"
          },
          "fetchedAt": {
            "format": "date-time",
            "title": "Fetchedat",
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          },
          "priceType": {
            "$ref": "#/components/schemas/PriceType"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "priceType",
          "value",
          "changeRate",
          "asOfDate",
          "asOfAt",
          "fetchedAt",
          "source",
          "freshness"
        ],
        "title": "LatestPriceResponse",
        "type": "object"
      },
...
      "Market": {
        "description": "表示资产所属市场，允许后续只新增枚举值。",
        "enum": [
          "CN"
        ],
        "title": "Market",
        "type": "string"
      },
...
      "PriceType": {
        "description": "严格区分股票价格、官方单位净值和盘中估算净值。",
        "enum": [
          "STOCK_LAST",
          "FUND_OFFICIAL_NAV",
          "FUND_ESTIMATED_NAV"
        ],
        "title": "PriceType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI Pick Stock API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/ai-conversations/{conversation_id}": {
      "get": {
        "description": "读取属于当前用户的完整多轮对话。",
        "operationId": "get_ai_conversation_api_v1_ai_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ai Conversation",
        "tags": [
          "ai-conversations"
        ]
      }
    },
    "/api/v1/ai-conversations/{conversation_id}/messages/stream": {
      "post": {
        "description": "用 SSE 实时发送 Codex Skill 状态和逐段回复。",
        "operationId": "stream_ai_conversation_message_api_v1_ai_conversations__conversation_id__messages_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConversationMessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Codex Agent 增量事件流"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Ai Conversation Message",
        "tags": [
          "ai-conversations"
        ]
      }
    },
...
    "/api/v1/instruments": {
      "get": {
        "description": "按代码或名称分页搜索当前一期支持的股票和基金。",
        "operationId": "search_instruments_api_v1_instruments_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "assetType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssetType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assettype"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Instruments",
        "tags": [
          "instruments"
        ]
      }
    },
    "/api/v1/instruments/{instrument_id}": {
      "get": {
        "description": "返回指定一期资产及各价格口径的最新本地快照。",
        "operationId": "get_instrument_api_v1_instruments__instrument_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instrument Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Instrument",
        "tags": [
          "instruments"
        ]
      }
    },
    "/api/v1/instruments/{instrument_id}/ai-analysis": {
      "get": {
        "description": "读取最近报告；尚无报告返回空值，不产生新的模型费用。",
        "operationId": "get_ai_analysis_api_v1_instruments__instrument_id__ai_analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instrument Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AIAnalysisResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Ai Analysis Api V1 Instruments  Instrument Id  Ai Analysis Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ai Analysis",
        "tags": [
          "ai-analysis"
        ]
      },
      "post": {
        "description": "刷新真实历史数据并手动生成或覆盖最近报告。",
        "operationId": "generate_ai_analysis_api_v1_instruments__instrument_id__ai_analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instrument Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Ai Analysis",
        "tags": [
          "ai-analysis"
        ]
      }
    },
    "/api/v1/instruments/{instrument_id}/ai-conversations": {
      "post": {
        "description": "为持仓或自选中的标的创建一个新的 Codex Session。",
        "operationId": "create_ai_conversation_api_v1_instruments__instrument_id__ai_conversations_post",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instrument Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ai Conversation",
        "tags": [
          "ai-analysis"
        ]
      }
    },
    "/api/v1/instruments/{instrument_id}/ai-conversations/latest": {
      "get": {
        "description": "读取当前用户指定标的最近一次多轮会话。",
        "operationId": "get_latest_ai_conversation_api_v1_instruments__instrument_id__ai_conversations_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instrument Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AIConversationResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Latest Ai Conversation Api V1 Instruments  Instrument Id  Ai Conversations Latest Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Ai Conversation",
        "tags": [
          "ai-analysis"
        ]
      }
    },
...
  }
}